extracted common functions into separate files
